[BUG] ensure we don't close listener fds twice on errors
If a listener fails to stop cleanly, we have to shut it and mark
it as such, so that further attempts to close it will not cause
unpredictable results. A problem could be triggered on FreeBSD 6.2
during a reload where an attempt to close a corrupted fd was seen.
Thanks to Anders Nordby for the detailed bug report.